+++ source/influx/influx_graphics/source/influx_graphics/d3d12/dx12_descriptors.cpp
@@ -129,7 +129,7 @@ namespace influx::graphics
 		D3D12_GPU_DESCRIPTOR_HANDLE gpu_handle{};
 		gpu_handle.ptr = reinterpret_cast<SIZE_T>(handle);
 
-		return (uint32)(gpu_handle.ptr - gpu_base.ptr) / m_descriptor_stride;
+		return (uint32)(gpu_handle.ptr - gpu_base.ptr) / static_cast<uint32>(m_descriptor_stride);
 	}
 
 	uint32 dx12_descriptor_heap::cpu_handle_to_index(descriptor_handle handle) const
@@ -138,7 +138,7 @@ namespace influx::graphics
 		D3D12_CPU_DESCRIPTOR_HANDLE cpu_handle{};
 		cpu_handle.ptr = reinterpret_cast<SIZE_T>(handle);
 
-		return (uint32)(cpu_handle.ptr - cpu_base.ptr) / m_descriptor_stride;
+		return (uint32)(cpu_handle.ptr - cpu_base.ptr) / static_cast<uint32>(m_descriptor_stride);
 	}
 
 	descriptor_handle dx12_descriptor_heap::index_to_gpu_handle(uint32 index) const
